requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
pandas==2.2.0
openpyxl==3.1.2
pdfplumber==0.10.3
openai==1.12.0
python-dotenv==1.0.0 